% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_PRS.R
\name{MR_PRS}
\alias{MR_PRS}
\title{Multivariate Mendelian Randomization (MVMR) Analysis Using PRS}
\usage{
MR_PRS(
  outcomefile,
  exposure_list,
  prscsxpath,
  plinkpath,
  conda_env,
  ref_dir,
  bfile,
  indMR,
  intercept = F
)
}
\arguments{
\item{outcomefile}{Path to the outcome file containing GWAS summary statistics for the outcome of interest. It should include columns for SNP identifiers, chromosome index, base pair positions, effect alleles, other alleles, Z-scores, and P-values.}

\item{exposure_list}{A list of paths to exposure files. Exposure file corresponding to each path should have the same column as outcome file.}

\item{prscsxpath}{Path to PRSCSx directory}

\item{plinkpath}{Path to Plink software}

\item{conda_env}{The name of the environment of PRSCSx in conda}

\item{ref_dir}{The path of refenece panels used in PRSCSx}

\item{bfile}{A list of PLINK bed files used in prediction.}

\item{indMR}{The IDs of individual in UK Biobank used to perform MVMR analysis.}

\item{intercept}{If an intercept is included in the MVMR analysis.}

\item{Nvec}{A vector of sample sizes of outcome and exposures GWAS. The first element should be the sample size of outcome GWAS.}
}
\value{
An object of class lm representing the linear model fit
}
\description{
This function performs a comprehensive Multivariate Mendelian Randomization analysis. It starts by processing GWAS summary data for an outcome and multiple exposures. PRS-CSx is then utilized to estimate the Polygenic Risk Scores (PRS) based on these summary statistics and genotype data from the UK Biobank. Finally, it performs MVMR to investigate the causal effects of the exposures on the outcome.
}
