% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisMR_PRS.R
\name{cisMR_PRS}
\alias{cisMR_PRS}
\title{Cis Multivariate Mendelian Randomization (cisMVMR) Analysis Using PRS}
\usage{
cisMR_PRS(
  outcomefile,
  CHR,
  BPcenter,
  BPtol,
  eQTL_list,
  prscsxpath,
  plinkpath,
  conda_env,
  ref_dir,
  bfile,
  indMR,
  intercept = F
)
}
\arguments{
\item{outcomefile}{Path to the outcome file containing GWAS summary statistics for the outcome of interest. It should include columns for SNP identifiers, chromosome index, base pair positions, effect alleles, other alleles, Z-scores, and P-values.}

\item{CHR}{The chromosome of interest for the analysis.}

\item{BPcenter}{The central base pair position of the genomic region in the chromosome of interest.}

\item{BPtol}{A threshold for the furthest SNP from the central base pair, used to define the genomic region of interest.}

\item{eQTL_list}{A list of paths to exposure files. Exposure file corresponding to each path should have the same column as outcome file.}

\item{prscsxpath}{Path to PRSCSx directory}

\item{plinkpath}{Path to Plink software}

\item{conda_env}{The name of the environment of PRSCSx in conda}

\item{ref_dir}{The path of refenece panels used in PRSCSx}

\item{bfile}{The PLINK  bed files used in prediction.}

\item{indMR}{The IDs of individual in UK Biobank used to perform MVMR analysis.}

\item{intercept}{If an intercept is included in the MVMR analysis.}
}
\value{
An object of class lm representing the linear model fit
}
\description{
This function performs a comprehensive Cis Multivariate Mendelian Randomization analysis. It starts by processing GWAS summary data for an outcome and multiple exposures. PRS-CSx is then utilized to estimate the Polygenic Risk Scores (PRS) based on these summary statistics and genotype data from the UK Biobank. Finally, it performs MVMR to investigate the causal effects of the exposures on the outcome.
}
\examples{
MR_PRS(outcomefile = "path/to/outcomefile.csv",
       CHR = 1,
       BPcenter = 15000000,
       BPtol = 1500000,
       exposures_list = list("path/to/exposure1.csv", "path/to/exposure2.csv"),
       prscsxpath = "/path/to/PRScsx/",
       plinkpath = "/path/to/Plink/",
       conda_env = "prscsx_env",
       indMR = c(1, 2, 3, 4))
}
